moving ConfirmDropdown to its own folder; PublishOnDemandWidget.tsx: fixing publishStudioNote message removing tags;
